Trying to fix lint problems.
